requirement.txt
absl-py==2.0.0
astunparse==1.6.3
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
Flask==3.0.0
flatbuffers==23.5.26
gast==0.5.4
geographiclib==2.0
geopy==2.4.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.14.0
libclang==16.0.6
libretranslatepy==2.1.1
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mysql-connector-python==8.2.0
nltk==3.8.1
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2023.3.post1
qrcode==7.4.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tqdm==4.66.1
translate==3.6.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
webcolors==1.13
Werkzeug==3.0.1
wrapt==1.14.1
